Code cleanup (3)

+++ subprojects/griffon-javafx/src/test/groovy/griffon/javafx/beans/property/ResetableFloatPropertyTest.java
@@ -29,6 +29,6 @@ public class ResetableFloatPropertyTest extends AbstractResetablePropertyTestCas
     @Override
     @Nonnull
     protected Number getNewValue() {
-        return Float.valueOf(21f);
+        return 21f;
     }
 }
